pyproject.toml
[project]
name = "dir-to-docx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Joseph Abbey", email = "me@josephabbey.dev" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "pygments>=2.19.2",
    "python-docx>=1.2.0",
]

[project.scripts]
dir-to-docx = "dir_to_docx:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
